Implemented "chording-like" behavior (#40)
When clicking numbers with an equal number of flags, will click all unrevealed tiles. This is similar to the "chording" behavior from original minesweeper where you can click both left and right buttons on a revealed number, and if the number of flags equals, it will reveal the remaining tiles. Does not allow cheating. If you mis-mark tiles, it will "boom" on an error, again, just like original minesweeper. And if you have too many marked, will do nothing.

Because the event mechanism does not provide the location in the click handler, had to save the last non-revealed tile we were on in the on_pointer_moved function. My thought was it was better to get this feature in, than to overhaul the mouse events to make true "chording" be the trigger.